Solution to Find the Median of an Array

+++ Find the Median of an Array@@676c02002dbc8c3e7a280b8c/code.c
@@ -13,11 +13,10 @@ int main() {
         scanf("%d", &arr[i]);
     }
 
-    
+    // Bubble sort
     for (int i = 0; i < n - 1; i++) {
         for (int j = 0; j < n - i - 1; j++) {
             if (arr[j] > arr[j + 1]) {
-                
                 int temp = arr[j];
                 arr[j] = arr[j + 1];
                 arr[j + 1] = temp;
@@ -25,12 +24,12 @@ int main() {
         }
     }
 
-    
-    float median;
+    // Median calculation - only integer output
+    int median;
     if (n % 2 != 0) {
         median = arr[n / 2];
     } else {
-        median = (arr[(n - 1) / 2] + arr[n / 2]) / 2;
+        median = (arr[(n - 1) / 2] + arr[n / 2]) / 2;  // Integer division (no .0)
     }
 
     printf("%d\n", median);
